+++ src/modifier/velocity.rs
@@ -53,6 +53,7 @@ impl SetVelocityCircleModifier {
         context.make_fn(
             &func_name,
             "transform: mat4x4<f32>, particle: ptr<function, Particle>",
+            None,
             module,
             &mut |m: &mut Module, ctx: &mut dyn EvalContext| -> Result<String, ExprError> {
                 let center = ctx.eval(m, self.center)?;
@@ -196,6 +197,7 @@ impl SetVelocityTangentModifier {
         context.make_fn(
             &func_name,
             "transform: mat4x4<f32>, particle: ptr<function, Particle>",
+            None,
             module,
             &mut |m: &mut Module, ctx: &mut dyn EvalContext| -> Result<String, ExprError> {
                 let origin = ctx.eval(m, self.origin)?;